Intron::is_reliable() spread coverage in centered/symmetric fashion
The positions/bins of an intron are symmetric to reversal/swapping sign.
More precisely, for an intron with N positions, positions i and N - 1 -
i overlap the intron by the exact same number of basepairs on opposite
sides of the intron and are the same distance to/from the left vs right
exons. This motivates a desire for Intron::is_reliable()'s approach for
spreading nonzero coverage to be symmetric in this way. To do this, the
ext bins of spread coverage are done centered around the nonzero
readrate. When ext is even, this cannot be done exactly, so the extra
bin is spread outwards. Both directions can be outwards for bin N / 2
when N is odd. In this case, spread outwards to (ext + 1) bins.